requirements.txt
# MQTT client library
paho-mqtt==2.1.0

# Testing dependencies
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0

# Type checking (optional)
mypy==1.8.0
types-paho-mqtt==1.6.0.20240106